without scaling upd: print functions for all figures & VectorDoc

+++ kab8/Circumference.h
@@ -16,8 +16,8 @@ public:
 	double get_square();
 	double get_perimeter();
 	void move(char c, double k);
-	void rotate(unsigned short int arg);
-	void scale(unsigned short int arg);
+	void rotate(unsigned int arg);
+	void scale(unsigned int arg);
 	void set_centre(Point c);
 	double get_x();
 	double get_y();
